requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
docker==6.1.3
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
packaging==23.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
redis==4.6.0
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
